requirements.txt
fastapi
uvicorn[standard]
python-multipart
pymupdf
chromadb
google-generativeai
python-dotenv
numpy
scikit-learn
